adding robust tests
